Renamed 'AttachmentInstance' class to 'VolumeAttachment' in normalization with OpenStack API
